requirements.txt
# Webサーバーフレームワーク
fastapi

# ASGIサーバー (開発用・本番用)
uvicorn
gunicorn

# WebSocket通信
python-socketio

# Firebase連携
firebase-admin

# 環境変数管理
python-dotenv

# gemini CLI
httpx 
json

